Fix schema guess for so large a long integer value

+++ embulk-guess-csv/src/main/java/org/embulk/guess/csv/SchemaGuess.java
@@ -171,7 +171,7 @@ final class SchemaGuess {
         }
 
         try {
-            if (Integer.valueOf(str).toString().equals(str)) {
+            if (Long.valueOf(str).toString().equals(str)) {
                 return GuessedType.LONG;
             }
         } catch (final RuntimeException ex) {
